fr-BE and fr-CH Towords ToOrdinalWords eccentricities, second implementation
