+++ src/Humanizer/Humanizer.csproj
@@ -80,6 +80,9 @@
     <Compile Include="GrammaticalGender.cs" />
     <Compile Include="Localisation\GrammaticalNumber\RussianGrammaticalNumber.cs" />
     <Compile Include="Localisation\GrammaticalNumber\RussianGrammaticalNumberDetector.cs" />
+    <Compile Include="Localisation\NumberToWords\FrenchSwissNumberToWordsConverter.cs" />
+    <Compile Include="Localisation\NumberToWords\FrenchBelgianNumberToWordsConverter.cs" />
+    <Compile Include="Localisation\NumberToWords\FrenchAbstractNumberToWordsConverter.cs" />
     <Compile Include="Localisation\NumberToWords\NorwegianBokmalNumberToWordsConverter.cs" />
     <Compile Include="Localisation\NumberToWords\FinnishNumberToWordsConverter.cs" />
     <Compile Include="Localisation\NumberToWords\GenderedNumberToWordsConverter.cs" />
